third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/secureContexts/localhost-expected.txt

This is a testharness.js-based test.
PASS localhost is secure. 
Harness: the test ran to completion.
